casting removed from dataset, will be done in pipeline

+++ src/AvailableRooftopDataset.py
@@ -66,7 +66,7 @@ class AvailableRooftopDataset(Dataset):
         
         # Retrieve the image and transpose from (HxWxC) -> (CxHxW)
         image = io.imread(image_path)
-        image = image.astype(np.float32)
+        #image = image.astype(np.float32)
         image = image.transpose(2, 0, 1)
 
         label = np.zeros((1, 250, 250))
@@ -74,7 +74,8 @@ class AvailableRooftopDataset(Dataset):
             # Retrieve the label and transpose from (HxWxC) -> (CxHxW)
             label = io.imread(label_path, as_gray=True)
             label = label[np.newaxis,:]
-
+        
+        #label = label.astype(np.float32)
         sample = {'image': image, 'label': label}
         
         # Apply the transforms if any
